requirements.txt
gunicorn==23.0.0
alembic==1.14.0
python-dotenv==1.0.1
APScheduler==3.11.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-Cors==5.0.0
tabulate==0.9.0